requirements.txt
Flask
google-cloud-aiplatform
google-cloud-pubsub
Pillow
Flask-SocketIO